engine: an edge count reaches its guard at the width the caller counted it in
create_edge checks `arity > MAX_ARITY` and create_genesis_event fills a fixed
buffer, but both took the count as uint8_t and every caller in parallel_evolution
narrowed to that width first, so the guard read a value the caller no longer had.

An initial edge of 260 vertices arrived as arity 4, passed the MAX_ARITY check and
built a state from four vertices with 256 dropped; 256 vertices arrived as arity 0.
Arity 17..255 threw correctly, so the guard held everywhere except at multiples of
256. The device rejects the same input before narrowing (gpu/src/initial_upload.cu
checks tuple.size() against kMaxArity), so host and device disagreed on it.

On the genesis path the same narrowing made a 256-edge initial state present as
zero edges: create_genesis_event then registered a producer for none of them, and
every causal edge originating at genesis was absent from the run with no warning
and no counter. A 300-edge state registered 44.

Both parameters are now size_t and both guards read the full count. MAX_ARITY is
16, so the arity narrows to uint8_t after the check as before. A genesis event
produces every initial edge and Event::num_produced is one byte, so 255 is a
structural bound rather than a storage detail: it is named MAX_GENESIS_EDGES,
sizes the key buffer, and a state past it is refused with a message saying so
rather than served wrong. The FFI already reports std::exception to the caller.

Gates: all_tests 305/305 (1 paclet test skipped, no Wolfram on this box) in
build_prof, HG_ENGINE_STATS=ON. Two tests added beside the existing bounds set:
an arity of 256 and of 260 must throw, and a genesis event one edge past
MAX_GENESIS_EDGES must throw while exactly MAX_GENESIS_EDGES does not. Both fail
against the byte-wide parameters, which accept those counts silently.

+++ hypergraph/src/parallel_evolution.cpp
@@ -92,7 +92,7 @@ void ParallelEvolutionEngine::evolve(
     // Create initial state
     std::vector<EdgeId> edge_ids;
     for (const auto& edge : initial_edges) {
-        EdgeId eid = hg_->create_edge(edge.data(), static_cast<uint8_t>(edge.size()));
+        EdgeId eid = hg_->create_edge(edge.data(), edge.size());
         edge_ids.push_back(eid);
 
         // Track max vertex ID to ensure fresh vertices don't collide
@@ -143,7 +143,7 @@ void ParallelEvolutionEngine::evolve(
         [[maybe_unused]] EventId genesis_event = hg_->create_genesis_event(
             raw_state,
             edge_ids.data(),
-            static_cast<uint8_t>(edge_ids.size())
+            edge_ids.size()
         );
 
         // Emit visualization event for the genesis event
@@ -156,7 +156,7 @@ void ParallelEvolutionEngine::evolve(
             genesis_event,  // event_id (raw)
             genesis_event,  // canonical_event_id (same as raw for genesis)
             0,              // destroyed edges (none)
-            static_cast<uint8_t>(edge_ids.size())  // created edges
+            edge_ids.size()  // created edges
         );
 #endif
     }
@@ -278,7 +278,7 @@ StateId ParallelEvolutionEngine::create_initial_state_only(
     // Create edges
     std::vector<EdgeId> edge_ids;
     for (const auto& edge : edges) {
-        EdgeId eid = hg_->create_edge(edge.data(), static_cast<uint8_t>(edge.size()));
+        EdgeId eid = hg_->create_edge(edge.data(), edge.size());
         edge_ids.push_back(eid);
         for (VertexId v : edge) {
             hg_->reserve_vertices(v);
@@ -296,7 +296,7 @@ StateId ParallelEvolutionEngine::create_initial_state_only(
 
     // Create genesis event if enabled
     if (enable_genesis_events_) {
-        hg_->create_genesis_event(raw_state, edge_ids.data(), static_cast<uint8_t>(edge_ids.size()));
+        hg_->create_genesis_event(raw_state, edge_ids.data(), edge_ids.size());
     }
 
     // Mark as seen but do NOT submit match task
@@ -311,7 +311,7 @@ StateId ParallelEvolutionEngine::create_and_register_initial_state(
     // Create edges
     std::vector<EdgeId> edge_ids;
     for (const auto& edge : edges) {
-        EdgeId eid = hg_->create_edge(edge.data(), static_cast<uint8_t>(edge.size()));
+        EdgeId eid = hg_->create_edge(edge.data(), edge.size());
         edge_ids.push_back(eid);
         for (VertexId v : edge) {
             hg_->reserve_vertices(v);
@@ -329,7 +329,7 @@ StateId ParallelEvolutionEngine::create_and_register_initial_state(
 
     // Create genesis event if enabled
     if (enable_genesis_events_) {
-        hg_->create_genesis_event(raw_state, edge_ids.data(), static_cast<uint8_t>(edge_ids.size()));
+        hg_->create_genesis_event(raw_state, edge_ids.data(), edge_ids.size());
     }
 
     // Seed the quotient causal reconstruction at this root (depth 0). The reconstruction runs
